backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi[standard]>=0.118.0",
    "pydantic-settings>=2.11.0",
    "pymongo>=4.15.2",
    "python-dotenv>=1.1.1",
    "rapidfuzz>=3.14.1",
    "transformers[torch]>=4.56.2",
]
